Revert "6주차"
