+++ src/main/java/com/umc/umcserver/dto/CreatePostsReqDto.java
@@ -1,18 +1,18 @@
-package com.umc.umcserver.src.post.dto;
+package com.umc.umcserver.dto;
 
-import com.umc.umcserver.src.post.repository.Board;
+import com.umc.umcserver.repository.Board;
 import lombok.Getter;
-import lombok.RequiredArgsConstructor;
+import lombok.NoArgsConstructor;
 
 @Getter
-@RequiredArgsConstructor
-public class UpdatePostsReqDto {
+@NoArgsConstructor
+public class CreatePostsReqDto {
     private String title;
     private String content;
     private String writer;
 
     // Dto에서 필요한 부분 entity화 시킴킴
-    public Board toEntity(){
+   public Board toEntity(){
         return Board.builder()
                 .title(title)
                 .content(content)
