+++ src/main/java/com/umc/umcserver/repository/Board.java
@@ -1,4 +1,4 @@
-package com.umc.umcserver.src.post.repository;
+package com.umc.umcserver.repository;
 
 import lombok.Builder;
 import lombok.Getter;
@@ -36,16 +36,6 @@ public class Board {
     public void update(String title, String content, String writer){
         this.title = title;
         this.content = content;
-        this.writer = writer;
-    }
-
-    public void updateTitle(String title){
-        this.title = title;
-    }
-    public void updateContent(String content){
-        this.content = content;
-    }
-    public void updateWriter(String writer){
         this.writer = writer;
     }
 }
