+++ src/main/java/com/umc/umcserver/service/BoardService.java
@@ -1,21 +1,18 @@
-package com.umc.umcserver.src.post.service;
+package com.umc.umcserver.service;
 
-import com.umc.umcserver.src.post.dto.*;
-import com.umc.umcserver.src.post.repository.Board;
-import com.umc.umcserver.src.post.repository.BoardRepository;
+import com.umc.umcserver.dto.*;
+import com.umc.umcserver.repository.Board;
+import com.umc.umcserver.repository.BoardRepository;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
-import java.util.ArrayList;
 import java.util.List;
 import java.util.stream.Collectors;
 
 @RequiredArgsConstructor
 @Service
-@Transactional(readOnly = true)
 public class BoardService {
-
     private final BoardRepository boardRepository;
 
     // 게시물 생성
@@ -49,33 +46,14 @@ public class BoardService {
         return new UpdatePostsResDto(board);
     }
 
-    // 일부 정보 업데이트
-    @Transactional
-    public UpdatePostsResDto PatchBoard(Long postId, UpdatePostsReqDto requestDto){
-        Board board = boardRepository.findById(postId)
-                .orElseThrow(() -> new IllegalArgumentException("해당 id는 존재하지 않습니다"));
-        List<UpdatePostsReqDto> lists = new ArrayList<>();
-        lists.add(requestDto);
-
-        for (UpdatePostsReqDto list : lists) { // 업데이트 사항인것만 반영해주기
-            if (list.getTitle() != null)
-                board.updateTitle(list.getTitle());
-            if (list.getWriter() != null)
-                board.updateWriter(list.getWriter());
-            if (list.getContent() != null)
-                board.updateContent(list.getContent());
-        }
-
-        return new UpdatePostsResDto(board);
-    }
-
     // 게시글 제거
     @Transactional
-    public void deleteBoard(Long postId) {
+    public void deleteBoard(Long postId){
         Board board = boardRepository.findById(postId)
                 .orElseThrow(() -> new IllegalArgumentException("해당 id는 존재하지 않습니다"));
         boardRepository.delete(board);
     }
+
 
 
 }
